Move has chain account validation to account module
