+++ feature-staking-impl/src/main/java/io/novafoundation/nova/feature_staking_impl/domain/parachainStaking/main/welcome/ParachainStakingWelcomeValidationFailure.kt
@@ -1,7 +1,7 @@
 package io.novafoundation.nova.feature_staking_impl.domain.parachainStaking.main.welcome
 
 import io.novafoundation.nova.feature_account_api.domain.model.MetaAccount
-import io.novafoundation.nova.feature_staking_impl.domain.common.validation.NoChainAccountFoundError
+import io.novafoundation.nova.feature_account_api.domain.validation.NoChainAccountFoundError
 import io.novafoundation.nova.runtime.multiNetwork.chain.model.Chain
 
 sealed class ParachainStakingWelcomeValidationFailure {
